refactor code to match eslint
