+++ log.js
@@ -65,6 +65,8 @@ function getCallerFile() {
 
 			if (currentfile !== callerfile) return path.basename(callerfile);
 		}
+		//TODO
+		// eslint-disable-next-line no-empty
 	} catch (err) {}
 	return undefined;
 }
